Creates new basic project structure using typescript from scratch
In this commit, we once again restructure how we are going to structure the project.
This way differs in that we start more from the ground up, and closer adhere to the
project structure that we had previously defined. Within this layout, we will also
define how we chose to handle all of mongo datatypes with typescript