Moves server.js to the root directory of the project
In this commit we move main server.js file to the root directory. I was
initially hesitant on this, but think it is the lesser of two evils. Yes,
this script is technically the server executable so should be located
within there. However, by moving it to the root directory we no longer