Prototypes User Models for Mongo and tidies project routing.
In this commit we make two significant changes, (sadly I forgot to commit
previous changes and they got too intertwined :/). Firstly, we begin to
implement how we are going to integrate passport with our mongodb instance.
For this, we have defined a new schema for the user account and extract the
DB logic into its own file.

Secondly in this commit, we reformat how we handle our routing to make the
project less cluttered and more dynamic. We are now using a layered routing
approach.