Separates express logic from shopCartController for better coherence.
In this commit, we move the express router logic out of the
shopCartController. Our justification for this is it made understanding the
routers more complicated. Furthermore, our security middleware injects the
user into the req body. Without this middleware then there would be no user
context. This level of assumed dependence could lead to unforeseen bugs down
the track.