Adds get all items functionality to carts.
In this commit, we add a new path to the api that
allows us to view all information about each item
within a given shopping cat. This can be accessed
via '/api/v1/cart/get/items'. Within the response
we give a form of pseudo objects that also contain
their respective quantities.