Adds search for item api path and refactors shopListingController into shopServerRoutes
In this commit we make two notable changes. The first of which is we add the new api
path `/api/v1/shop/items/:itemId` where `itemId` is a request parameter containing the
MongoDB item id for the desired item. Within this commit, we also refactor some of our
logic previously stored within shopListingController into shopServerRoutes. The
justification for this change is that although the single callback function was
visibly tidier. The callback was making assumptions about the request such as whether
or not middleware was present and what request parameter / body is required. This
approach keeps the controller logic more isolated, as well as look more cohesive and
consistent with the rest of the application.