Adds place recommendations via the Google Places API.
In this commit, we add place recommendations by extracting the IP of
the request via Express. If this cannot be located (or we are on
localhost) then default to an arbitrary Wellington IP address. From
here we use the Google Places API to recommend a list of different
locations that match a computer store query. From here the data is
parsed and returned within a response. The mapping for this can be
accessed via '/api/v1/recommend/nearby'.