Refactors recommendation logic into recommendController.ts
In this commit, we refactor the previous commit into different pieces
of logic. Here the external facing logic is refactored into the
recommendController.ts whereas the extracting IP address logic is
placed within its own function within the recommendationsServerRoutes
The justification for this is it is purely mapping / routing logic
based and doesn't perform any additional business logic such as
interface with external or internal data sources.