Adds Google OAuth2 support for logging in via Google account.
In this commit, we add two new API routing and a new middleware to allow users to
login with their Google Account. The first api route '/api/v1/login/google' takes
the user to a select Google account screen, upon selection, the user is then taken
to the '/api/v1/login/google/callback' URl where the OAuth2 profile is passed via
the URL. From here, on the callback we can then create / find a user account
associated with the OAuth2 profile. This allows us to then generate a Jwt token
such that they can authenticate just like a user who signed in with email +
password.