Configures both testing and development MongoDbs.
In this commit, we make a couple notable changes. The first of which
is we configure `config.ts` to allow for both a development and a
testing mongodb atlas. This means that our testing db will be fully
isolated during testing. We also add an `assertType.ts`, that
contains a TypeScript function that we will likely use to verify the
responses from the server within our Mocha + Chai testing. Using
this function, we should be able to dynamically check the responses
by comparing them (by types, not values) with a pre-made expected
response.