Adds improved error handling for errors within thrown response bodies from Mongoose.
In this commit, we address a flaw with the implementation of the backend server routes.
Here, with bad inputs Mongoose was throwing an error (instead of returning null), this
meant that the response sent to the client was very messy and hard to extract the
information. Instead, here we add a function to extract two different kinds of messages
from errors from the server.