Adds user authentication tests and timeout to the server
In this commit we add a new 'Authentication' test suite to the tests.ts
file, as well as introduce some extra shop listing tests to handle invalid
cases. Within this commit, we also add a global 7s timeout to the entire
api routes for the server to disconnect if an active connection lasts that
long.