Implements reset password backend using email reset links.
In this commit, we implement password reset functionality, by
adding two new API calls to the backend. This allows us to
generate a new reset link that is then emailed to the users
inbox. From here, by using this token, the user is then able
to provide a new password (via backend) and the corresponding
user account will be updated.