Implements update account API path and fixes admin=false bug.
In this commit, we quickly implement the update account api
route. Within here, for security reasons we only expose the
displayName attribute to be updated. (Password also should
be handled separately). We also identify and resolve the bug
where any changes to the users account will result in the
user being set admin=false regardless.